fix: separate common commands
