+++ commands/commands.py
@@ -6,7 +6,7 @@ from commands.BlueBotCommands import BlueBotCommands
 from commands.ContextMenu import ContextMenu
 from commands.HashingCommands import HashingCommands
 from commands.SubCommand import SubCommand
-from commands.OptionCommands import OptionCommand
+from commands.EmbedCommands import EmbedCommands
 
 
 class Commands:
@@ -21,7 +21,7 @@ class Commands:
         HashingCommands(client=client, tree=tree, servers=servers)
         ContextMenu(client=client, tree=tree, servers=servers)
         BlueBotCommands(client=client, tree=tree, servers=servers)
-        OptionCommand(client=client, tree=tree, servers=servers)
+        EmbedCommands(client=client, tree=tree, servers=servers)
         # SubCommand(client=client, tree=tree, servers=servers)
 
     def register_Commands(self):
